requirements.txt
aiohttp==3.9.2
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.14.0
beautifulsoup4==4.12.3
blinker==1.6.2
Brotli==1.1.0
carehare==1.0.2
certifi==2023.7.22
cffi==1.15.1
channels==4.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cssselect2==0.7.0
distlib==0.3.6
distro==1.8.0
Django==4.2
django-4-jet==1.0.9
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==4.4.0
django-jazzmin==2.6.0
django-jet==1.0.8
django-jet-4==0.0.1
django-localflavor==4.0
django-model-utils==4.5.1
django-notifications-hq==1.8.3
django-select2==8.2.0
django-tinymce==4.0.0
django-widget-tweaks==1.5.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.1
dnspython==2.6.0
docopt==0.6.2
docxcompose==1.4.0
docxtpl==0.16.8
email-validator==1.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
filelock==3.12.0
Flask==2.3.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.53.1
frozenlist==1.3.3
greenlet==3.0.3
gunicorn==20.1.0
h11==0.14.0
html2text==2024.2.26
html5lib==1.1
htmlmin==0.1.12
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
jsmin==3.0.1
jsonfield==3.1.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lesscpy==0.15.1
lxml==4.9.2
Mako==1.3.2
Markdown==3.6
MarkupSafe==2.1.3
msgpack==1.0.8
multidict==6.0.4
mysqlclient==2.2.4
num2words==0.5.13
numpy==1.25.0
openai==0.27.6
opencv-python==4.10.0.82
openpyxl==3.1.5
outcome==1.2.0
packaging==23.1
pamqp==3.3.0
pandas==2.0.1
Pillow==10.0.1
platformdirs==3.5.1
ply==3.11
pycparser==2.21
pydyf==0.11.0
pymongo==4.6.1
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pypandoc==1.13
pyphen==0.16.0
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
python-stdnum==1.20
pytz==2023.3
rcssmin==1.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
selenium==4.12.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==1.4.29
sqlparse==0.4.4
swapper==1.3.0
tabula-py==2.7.0
tinycss2==1.3.0
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.23.0
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.0.0
xxhash==3.4.1
yarl==1.9.2
zopfli==0.2.3
